Fine

+++ src/Action/ContactController.php
@@ -46,10 +46,18 @@ class ContactController
         $args = (array)$request->getAttributes();
         $contact = $this->contactRepository->getContactById($args['id']);
 
-        if ($request->hasHeader('HX-Request')) {
-            return $this->view->render($response, 'partial/show_contact.twig', ['contact' => $contact]);
+        if ($contact) {
+            if ($request->hasHeader('HX-Request')) {
+                return $this->view->render($response, 'partial/show_contact.twig', ['contact' => $contact]);
+            } else {
+                return $this->view->render($response, 'full/show_contact.twig', ['contact' => $contact]);
+            }
         } else {
-            return $this->view->render($response, 'full/show_contact.twig', ['contact' => $contact]);
+            $this->c->get('flash')->addMessage('status', 'No Contact Found');
+
+            return $response
+                ->withHeader('Location', '/contacts')
+                ->withStatus(302);
         }
     }
 
@@ -58,10 +66,18 @@ class ContactController
         $args = (array)$request->getAttributes();
         $contact = $this->contactRepository->getContactById($args['id']);
 
-        if ($request->hasHeader('HX-Request')) {
-            return $this->view->render($response, 'partial/edit_contact.twig', ['contact' => $contact]);
+        if ($contact) {
+            if ($request->hasHeader('HX-Request')) {
+                return $this->view->render($response, 'partial/edit_contact.twig', ['contact' => $contact]);
+            } else {
+                return $this->view->render($response, 'full/edit_contact.twig', ['contact' => $contact]);
+            }
         } else {
-            return $this->view->render($response, 'full/edit_contact.twig', ['contact' => $contact]);
+            $this->c->get('flash')->addMessage('status', 'No Contact Found');
+
+            return $response
+                ->withHeader('Location', '/contacts')
+                ->withStatus(302);
         }
     }
 
@@ -70,9 +86,12 @@ class ContactController
         $args = (array)$request->getAttributes();
         $contact = $this->contactRepository->getContactById($args['id']);
 
-        $this->contactRepository->deleteContact($contact);
-
-        $this->c->get('flash')->addMessage('status', 'Deleted Contact!');
+        if ($contact) {
+            $this->contactRepository->deleteContact($contact);
+            $this->c->get('flash')->addMessage('status', 'Deleted Contact!');
+        } else {
+            $this->c->get('flash')->addMessage('status', 'No Contact Found');
+        }
 
         return $response
             ->withHeader('Location', '/contacts')
@@ -100,9 +119,36 @@ class ContactController
         $contact->setPhone($params['phone']);
         $contact->setRegisteredAt();
 
-        $this->contactRepository->setNewContact($contact);
+        $this->contactRepository->setContact($contact);
 
         $this->c->get('flash')->addMessage('status', 'Created New Contact!');
+
+        return $response
+            ->withHeader('Location', '/contacts')
+            ->withStatus(302);
+    }
+
+    public function update(Request $request, Response $response): Response
+    {
+        $args = (array)$request->getAttributes();
+
+        $contact = $this->contactRepository->getContactById($args['id']);
+
+        $params = (array)$request->getParsedBody();
+
+        if ($contact) {
+            $contact->setFirstName($params['first_name']);
+            $contact->setLastName($params['last_name']);
+            $contact->setEmail($params['email']);
+            $contact->setPhone($params['phone']);
+
+            $this->contactRepository->setContact($contact);
+
+            $this->c->get('flash')->addMessage('status', 'Updated Contact!');
+        } else {
+
+            $this->c->get('flash')->addMessage('status', 'No Contact Found');
+        }
 
         return $response
             ->withHeader('Location', '/contacts')
